General code clean-up.
